refactor: Move UserAuth to src/types
