+++ packages/api/src/constants/index.ts
@@ -2,3 +2,5 @@ export const ISSUER = 'FELABS'
 export const TIME = {
   '1day': 1000 * 60 * 60 * 24,
 } as const
+export const TOKEN_COOKEY_NAME = 'tdl_token'
+export const APP_SECRET = process.env.APP_SECRET as string
