--- packages/api/src/user/user.interface.ts
@@ -1,6 +1,0 @@
-export interface UserAuth {
-  exp: number // token expired at (epoch datetime)
-  iat: number // token issued at (epoch datetime)
-  iss: string // issuer
-  uid: string // user id
-}
